fix a bug in shared heap malloc when it's memory64

+++ core/iwasm/common/wasm_memory.c
@@ -796,10 +796,9 @@ wasm_runtime_shared_heap_malloc(WASMModuleInstanceCommon *module_inst,
         *p_native_addr = native_addr;
     }
 
-    return memory->is_memory64
-               ? shared_heap->start_off_mem64
-               : shared_heap->start_off_mem32
-                     + ((uint8 *)native_addr - shared_heap->base_addr);
+    return (memory->is_memory64 ? shared_heap->start_off_mem64
+                                : shared_heap->start_off_mem32)
+           + (uintptr_t)((uint8 *)native_addr - shared_heap->base_addr);
 }
 
 void
